requirements.txt
apispec==5.1.1
apispec_webframeworks==0.5.2
Flask==1.1.2
Flask_SQLAlchemy==2.4.3
flask_swagger_ui==3.36.0
Flask_User==1.0.2.2
PyJWT==2.3.0
Werkzeug==1.0.1
